chore: Add backend config to schema
## Motivation

While using the [DevTools
plugin](https://github.com/backstage/backstage/blob/master/plugins/devtools/README.md)
we noticed that the `clientSecret` was being exposed in plain text.

By marking its visibility we can get it to be redacted in that config
view.

## Summary

Added the missing configuration fields to the `config.d.ts` file. I'm
not sure if they should be in this file in the frontend plugin or if you
prefer that I create a new file in the backend plugin instead. Happy to
do the changes if needed.


## Release Notes

Ensure that the config fields for the credentials have the correct
visibility and are redacted in the Config view of the devTools plugin.